requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
dj-database-url==1.2.0
Django==4.1.6
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
fonttools==4.38.0
gunicorn==20.1.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
mlxtend==0.21.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.7
whitenoise==6.3.0
